update encrypt & decrypt
